% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{Arich}
\alias{Arich}
\title{A Arich function}
\usage{
Arich(reads = NULL, bsgenome = NULL)
}
\arguments{
\item{reads}{A data frame containing alignment results and identified polyA sites.}

\item{bsgenome}{A BSgenome object for reference genome. e.g. 'BSgenome.Hsapiens.UCSC.hg38'.}
}
\value{
A data frame containing information about polyA sites near A-rich regions
}
\description{
This function to search A-rich fragments surrounding polyA sites.
}
