% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{cal.chisq}
\alias{cal.chisq}
\title{Calculate a chi-square metric}
\usage{
cal.chisq(fafile = NULL, ref.data = NULL, iteration = 100, use.size = 5000)
}
\arguments{
\item{fafile}{fasta file name, which contains upstream and downstream 100 bp sequences of PACs}

\item{ref.data}{Single nucleotide frequency distribution of the reference PACs.}

\item{iteration}{The number of random sampling iterations, default is 100.}

\item{use.size}{The number of PACs randomly sampled each time, default is 5000.}
}
\value{
Chi-Square (χ2) Statistic
}
\description{
to assess the similarity between the single nucleotide profile of
identified poly A sites and the reference profile
}
