% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{correctSNP}
\alias{correctSNP}
\title{A correctSNP  function}
\usage{
correctSNP(reads = NULL)
}
\arguments{
\item{reads}{A data frame containing alignment results.}
}
\value{
A data frame containing corrected polyA sites affected by SNPs
}
\description{
This function to detect potential polyA sites affected by SNPs.
}
