% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{data.PACds}
\alias{data.PACds}
\title{data.PACds: Example annotaed PACs Dataset for Human}
\format{
A PACdataset object.
}
\source{
Generated from demo data
}
\usage{
data.PACds
}
\description{
A PACdataset containing example polyA site clusters (PACs) information for Human.
}
\examples{
data(data.PACds)
head(data.PACds@anno)
}
\keyword{datasets}
