% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{findPeaks}
\alias{findPeaks}
\title{A findPeaks function}
\usage{
findPeaks(sub_pos,sub_wts,min_delta=24)
}
\arguments{
\item{sub_pos}{The genomic coordinates of poly(A) sites.}

\item{sub_wts}{The numbers of reads support each poly(A) site in 'sub_pos'.}

\item{min_delta}{A cutoff to limit the minimum distance between two centers.}
}
\value{
A tibble object with five columns 'group', 'start', 'end', 'sum.wts', 'center'.
}
\description{
Cluster the one-dimensional poly(A) sites into groups based on local density
}
