% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{FindPTA}
\alias{FindPTA}
\title{A FindPTA function}
\usage{
FindPTA(
  bam = NULL,
  yieldSize = 2.6e+07,
  reverse = FALSE,
  bsgenome = NULL,
  d = 24,
  poly = "T",
  adjust.chr = FALSE,
  threeUTRregion = NULL,
  cutoffCount = 5,
  ext3UTRlen = 1000,
  isDRS = FALSE,
  run.quantify = TRUE
)
}
\arguments{
\item{bam}{The file name of the 'BAM' file to be processed.}

\item{yieldSize}{Number of records to yield each time the file is read from
with Rsamtools::scanBam.}

\item{reverse}{If TRUE, reverse complement the reads from the BAM file.
BAM files store reads mapping to the minus strand as though they
are on the plus strand. Default is FALSE.}

\item{bsgenome}{A BSgenome object for reference genome. e.g. 'BSgenome.Hsapiens.UCSC.hg38'.}

\item{d}{A distance to group nearby polyA sites into a polyA site cluster (PAC). Default is 24 nt.}

\item{poly}{Specifies the type of polyA tail at the 3' end of the sequence.
Must be either 'T' (for a poly(T) stretch, e.g., TTTTTT-like tails)
or 'A' (for a poly(A) stretch, e.g., AAAAAA-like tails).
Default is 'T'.}

\item{adjust.chr}{Set to TRUE to add 'chr' prefix if input chromosome names
in BAM file lack 'chr' prefix but the genome sequence includes it.
Default is FALSE.}

\item{threeUTRregion}{A GRanges object for 3'UTR region, will use level-8 reads in identification of polyA sites.
Recommended to use if input reads lack polyA tails.}

\item{cutoffCount}{Minimum count threshold for PACs identified at the V8 level
to be used for polyA determination.}

\item{ext3UTRlen}{An integer denoting the extended length of 3'UTR region. Default is 1000 nt.}

\item{isDRS}{Set to TRUE if input is Direct RNA Sequencing (DRS) data.
Default is FALSE.}

\item{run.quantify}{If TRUE (default), quantify polyA sites.}
}
\value{
A list containing two elements about alignment results and PACs.
\itemize{
  \item \code{pa.table}: A \code{data.frame} with the following columns:
    \describe{
      \item{\code{\bold{readName}}}{Name of the read.}
      \item{\code{\bold{flag}}}{Bitwise FLAG.}
      \item{\code{\bold{cigar}}}{CIGAR string.}
      \item{\code{\bold{chr}}}{The name of the chromosome.}
      \item{\code{\bold{start}}}{The starting position of the feature in the chromosome.}
      \item{\code{\bold{end}}}{The ending position of the feature in the chromosome.}
      \item{\code{\bold{strand}}}{The strand direction of the feature in the chromosome.}
      \item{\code{\bold{coord}}}{The coordinate of the potential polyA site.}
      \item{\code{\bold{seq}}}{The sequence of the input trimmed read.}
      \item{\code{\bold{softClipFragment}}}{The fragment of the soft-clipping.}
      \item{\code{\bold{trimmed_seq}}}{The trimmed polyA tail.}
      \item{\code{\bold{unmapped_seq}}}{The sequence in the original read that cannot be aligned to the reference genome.}
      \item{\code{\bold{reference_seq}}}{The reference sequence around the potential polyA site.}
      \item{\code{\bold{ref_match_trimmed}}}{The part of the reference sequence corresponding to \code{trimmed_seq}.}
      \item{\code{\bold{ref_match_unmapped}}}{The part of the reference sequence corresponding to \code{unmapped_seq}.}
      \item{\code{\bold{is_Arich}}}{Whether it is an A-rich polyA site.}
      \item{\code{\bold{A_number}}}{The total number of adenines (A) within the reference sequence, spanning 10 bp upstream and downstream of the polyA site.}
      \item{\code{\bold{diff_a}}}{The total number of non-genomic As at the 3’ end.}
      \item{\code{\bold{level}}}{The subclass or confidence level of the aligned read.}
      \item{\code{\bold{class}}}{The category of the aligned read.}
      \item{\code{\bold{use.as.count}}}{Whether the read is used in the quantification of PACs.}
      \item{\code{\bold{PAid}}}{The identity of PACs.}
    }
  \item \code{pa.coord}: A \code{data.frame} with the following columns:
    \describe{
      \item{\code{\bold{seqnames}}}{The name of the chromosome of PACs.}
      \item{\code{\bold{start}}}{The starting position of PACs in the chromosome.}
      \item{\code{\bold{end}}}{The ending position of PACs in the chromosome.}
      \item{\code{\bold{width}}}{The width of PACs.}
      \item{\code{\bold{coord}}}{The representative coordinate of PACs.}
      \item{\code{\bold{coord_readName}}}{The name of the read supporting the representative polyA site.}
      \item{\code{\bold{coord_unmappedA}}}{The total number of non-genomic As at the 3’ end of the representative read.}
      \item{\code{\bold{coord_level}}}{The confidence level of the representative polyA site.}
      \item{\code{\bold{UPA_coord}}}{The coordinates of aligned reads in corresponding PACs.}
      \item{\code{\bold{UPA_readName}}}{The name of aligned reads in corresponding PACs.}
      \item{\code{\bold{UPA_level}}}{The confidence level label of aligned reads in corresponding PACs.}
      \item{\code{\bold{PAid}}}{The identity of PACs.}
      \item{\code{\bold{total.count}}}{The total counts of PACs based on all categories of aligned reads.}
      \item{\code{\bold{count_with.polyA}}}{The total counts of PACs based on C1-C3 of aligned reads.}
    }
}
}
\description{
This function allows you to identify and quantify polyA sites from different
3'seq data.
}
