% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{findTailAT}
\alias{findTailAT}
\title{A findTailAT function}
\usage{
findTailAT(
  infile = NULL,
  poly,
  ml = 20,
  mg = 5,
  mp = 8,
  mtail = 8,
  mper = 0.75,
  mm = 2,
  mr = 3,
  reg = 1,
  bar = 0,
  deep = TRUE,
  review = TRUE,
  debug = TRUE,
  odir = NULL,
  suf = NULL
)
}
\arguments{
\item{infile}{input fastq file.}

\item{poly}{should be one of "A", "T"}

\item{ml}{min length after trim. Default is 20.}

\item{mg}{margin from the start (poly=T) or to the end (poly=A) (default is 5)}

\item{mp}{min length of successive poly (default is 8)}

\item{mtail}{min length of trimmed tail (default is 8)}

\item{mper}{min percent of 'A\\/T' in trimmed tail (default is 0.75)}

\item{mm}{mismatch between TxxTTT (default is 2)}

\item{mr}{minT in reg (default is 3)}

\item{reg}{regexp string, default is 1 for loose searching, or 2 for stricter searching.
1 = \code{'^.{0,mg}?(T{mr,}[^T]{0,mm})*T{mp,}([^T]{0,mm}T{mr,})*'}
2 = \code{'^.{0,mg}?(T{mp,}([^T]{0,mm}T{mr,})*)'}.}

\item{bar}{the last bast position of three barcode, default is 0.}

\item{deep}{T(default) or FALSE, if T, then when no match was found for reg, will do deeper searching for tails like TTTTTCTTTTCTCTTTTTTTT..}

\item{review}{T (default) or F, if T then when no match was found for reg or deep, do deeper search for tails like AACCCCTTTTTTTTTTTTTTTTT...}

\item{debug}{T (default) or F, if T then output debugging files.}

\item{odir}{output path, if nULL then the same as infile dir.}

\item{suf}{suffix for output files, like <infile><.suf>.T/A.fq, default is NULL.}
}
\value{
character. Output several files:
  \itemize{
    \item "*.T.fq" - Contains reads with names annotated to indicate polyT stretches at their 3' ends.
    \item "*.A.fq" - Contains reads with names annotated to indicate polyA stretches at their 3' ends.
  }
These files can serve as input for alignment tools.
}
\description{
findTailAT finds and trims A or T tails from a fastq file.
}
\examples{
file_T <- system.file("extdata", "SRR1168402_T.fastq", package = "PolyAseqTrap")
findTailAT(infile=file_T, odir=NULL,
poly='T', ml=20, mp=5, mg=10, mm=2, deep=FALSE,
mtail=6, mper=0.75, mr=3, review=TRUE, debug=TRUE,
bar=0, reg=1, suf=NULL)

file_A <- system.file("extdata", "SRR11837378_A.fastq", package = "PolyAseqTrap")
findTailAT(infile=file_A, odir=NULL,
poly='A', ml=20, mp=5, mg=10, mm=2, deep=0, mtail=6, mper=0.75, mr=3,
review=TRUE, debug=TRUE, bar=0, reg=1, suf=NULL)
}
