% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{generateFASTA}
\alias{generateFASTA}
\title{A generateFASTA function}
\usage{
generateFASTA(reads = NULL, bsgenome = NULL, output.name = NULL)
}
\arguments{
\item{reads}{A data frame contains alignment result and identified A-rich polyA sites.}

\item{bsgenome}{A BSgenome object for reference genome. e.g. 'BSgenome.Hsapiens.UCSC.hg38'.}

\item{output.name}{The name of output fastq file.}
}
\value{
A fastq file
}
\description{
Extract sequence from upstream 100 nt to downstream 100 nt region of polyA sites.
}
