% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{PACs_human}
\alias{PACs_human}
\title{PACs_human: Example PACs Dataset for Human}
\format{
A list containing two elements about alignment results and PACs.
}
\source{
Generated from demo data
}
\usage{
PACs_human
}
\description{
A dataset containing example polyA site clusters (PACs) information for human.
}
\examples{
data(PACs_human)
head(PACs_human$pa.table)
head(PACs_human$pa.coord)
}
\keyword{datasets}
