% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{PACs_mouse}
\alias{PACs_mouse}
\title{PACs_mouse: Example PACs Dataset for Mouse}
\format{
A list containing two elements about alignment results and PACs.
}
\source{
Generated from demo data
}
\usage{
PACs_mouse
}
\description{
A dataset containing example polyA site clusters (PACs) information for Mouse.
}
\examples{
data(PACs_mouse)
head(PACs_mouse$pa.table)
head(PACs_mouse$pa.coord)
}
\keyword{datasets}
