% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{PACs_tair}
\alias{PACs_tair}
\title{PACs_tair: Example PACs Dataset for Arabidopsis}
\format{
A list containing two elements about alignment results and PACs.
}
\source{
Generated from demo data
}
\usage{
PACs_tair
}
\description{
A dataset containing example polyA site clusters (PACs) information for Arabidopsis.
}
\examples{
data(PACs_tair)
head(PACs_tair$pa.table)
head(PACs_tair$pa.coord)
}
\keyword{datasets}
