% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{plot_summary}
\alias{plot_summary}
\title{A summary report function}
\usage{
plot_summary(data = NULL)
}
\arguments{
\item{data}{A data frame inclues class, subclass, genomics features and polyA signals.}
}
\value{
plots
}
\description{
plot_summary provides an overview of PACs at different levels,
including their categories, subclasses, genomics region, and signal distribution
}
