% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{process_data}
\alias{process_data}
\title{A process_data function}
\usage{
process_data(aln = NULL, bsgenome = NULL, adjust.chr = FALSE, poly = NULL)
}
\arguments{
\item{aln}{A alignment result from bam file.}

\item{bsgenome}{A BSgenome object for reference genome. e.g. 'BSgenome.Hsapiens.UCSC.hg38'.}

\item{adjust.chr}{Set to TRUE to add 'chr' prefix if input chromosome names
in BAM file lack 'chr' prefix but the genome sequence includes it.
Default is FALSE.}

\item{poly}{Specifies the type of polyA tail at the 3' end of the sequence.
Must be either 'T' (for a poly(T) stretch, e.g., TTTTTT-like tails)
or 'A' (for a poly(A) stretch, e.g., AAAAAA-like tails).
Default is 'T'.}
}
\value{
A data frame containing alignment results, including chromosome, start, end,
        strand, trimmed read sequence, polyA tail sequence, and reference sequence
        around potential polyA sites.
}
\description{
This function processes a BAM file to extract alignment information and
polyA tail details.
}
