% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{process_tail_nopure_v1}
\alias{process_tail_nopure_v1}
\title{A process_tail_nopure_v1  function}
\usage{
process_tail_nopure_v1(aln.tail.nopure.v1 = NULL, aln.result = NULL)
}
\arguments{
\item{aln.tail.nopure.v1}{A alignment result of reads with perfectly matched polyA tails and soft-clipping.}

\item{aln.result}{A alignment result from bam file.}
}
\value{
A data frame containing more detailed alignment results.
}
\description{
This function to detect polyA sites from reads with perfectly matched polyA tails and soft-clipping.
}
