% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{process_tail_nopure_v3}
\alias{process_tail_nopure_v3}
\title{A process_tail_nopure_v3  function}
\usage{
process_tail_nopure_v3(aln.tail.nopure.v3 = NULL, aln.result = NULL)
}
\arguments{
\item{aln.tail.nopure.v3}{A alignment result of reads with partially matched polyA tails.}

\item{aln.result}{A alignment result from bam file.}
}
\value{
A data frame containing more detailed alignment results.
}
\description{
This function to detect polyA sites from reads with partially matched polyA tails.
}
