% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{process_tail_pure}
\alias{process_tail_pure}
\alias{process_tail_nopure_v4}
\title{A process_tail_pure function}
\usage{
process_tail_pure(aln.tail.pure = NULL)

process_tail_nopure_v4(aln.tail.nopure.v4 = NULL)
}
\arguments{
\item{aln.tail.pure}{A alignment result about reads with perfectly matched polyA tails.}

\item{aln.tail.nopure.v4}{A alignment result about reads with perfectly matched polyA tails.}
}
\value{
A data frame of identified polyA sites.

A data frame of alignment result.
}
\description{
This function to identify polyA sites from reads with perfectly matched polyA tails.

This function to exclude reads with perfectly matched polyA tails, but polyA tails
actually from reference genome
}
