% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{recycle_v8}
\alias{recycle_v8}
\title{A recycle_v8 function}
\usage{
recycle_v8(
  aln.result = NULL,
  cutoffCount = NULL,
  threeUTRregion = NULL,
  ext3UTRlen = 1000
)
}
\arguments{
\item{aln.result}{A data frame contains alignment results and identified polyA sites.}

\item{cutoffCount}{Minimum count threshold for PACs identified at the V8 level
to be used for polyA determination.}

\item{threeUTRregion}{A GRanges object for 3'UTR region, will use level-8 reads in identification of polyA sites.
Recommended to use if input reads lack polyA tails.}

\item{ext3UTRlen}{An integer denoting the extended length of 3'UTR region. Default is 1000 nt.}
}
\value{
Identifed V8 polyA sites.
}
\description{
This function to detect V8 polyA sites.
}
