% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{ref.data}
\alias{ref.data}
\title{ref.data: Example single nucleotide profile of reference PACs in human}
\format{
A data frame.
}
\source{
Generated from human reference PACs database
}
\usage{
ref.data
}
\description{
A referendce polyA site clusters (PACs) profile for Human.
}
\examples{
data(ref.data)
head(ref.data)
}
\keyword{datasets}
