% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{resut.PA}
\alias{resut.PA}
\title{A resut.PA function}
\usage{
resut.PA(aln.result = NULL, d = NULL)
}
\arguments{
\item{aln.result}{A data frame contains alignment results and identified polyA sites.}

\item{d}{A distance to group nearby polyA sites into a polyA site cluster (PAC). Default is 24 nt.}
}
\value{
A list containing two element. The 'pa.table' element
        is a mapped table of 3'reads, including sequence alignment results,
        base compositions of the polyA tail, polyA site position, and
        the confidence level of reads used to identify polyA sites (from V1 to V8).
        the 'pa.coord'  element contains identified and quantified PACs.
}
\description{
TThis function to classify PACs and quantify PACs.
}
