% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAseqTrap_funclib.R
\name{split_pac}
\alias{split_pac}
\title{A split_pac function}
\usage{
split_pac(pa.data = NULL, d = 24, mc.cores = 4)
}
\arguments{
\item{pa.data}{A list of polyA information, the output of the FindPTA function}

\item{d}{A cutoff limiting the max distance between two adjacent poly(A) sites in a poly(A) site cluster (PAC), default value is 24 nt.}

\item{mc.cores}{An integer indicating the number of processors/cores to perform the clustering, default value is 4.}
}
\value{
A list contains re-defined PACs based on a weighted density peak clustering method
}
\description{
Cluster poly(A) sites into groups by a certain distance iterative.
}
